requirements.txt
streamlit
openai
python-dotenv
reportlab
python-docx
requests